==> notebooks/02_spotify_kaggle_dt.ipynb <==
# Cell 1 - Import, percorsi, autenticazione Spotify

import os
import pandas as pd
import numpy as np

DATA_PROCESSED_DIR = os.path.join("..", "data", "processed")
os.makedirs(DATA_PROCESSED_DIR, exist_ok=True)

print("✓ Imports OK")

## Cell 2 - Song's Dataset

## Kaggle set

kaggle_path = "/Users/Vitaliano/Desktop/Data Science/FDS - Fundamentals of Data Science/Homework di Gruppo/Final Project/Music-mood pred/data/data.csv"

df= pd.read_csv(kaggle_path)
print("Kaggle dataset shape:", df.shape)
df.head()


rename_cols = {
    "id": "track_id",
    "name": "track_name",
    "artists": "artist_name",
    "popularity": "popularity",
    "release_date": "release_date"
}

df = df.rename(columns=rename_cols)

# colonne da mantenere
keep = [
    "track_id", "track_name", "artist_name", "popularity", "release_date",
    "acousticness","danceability","energy","instrumentalness","liveness",
    "loudness","speechiness","tempo","valence","duration_ms"
]

df = df[keep].dropna(subset=["track_id"])
df["source_type"] = "kaggle"
df["source_label"] = "spotify_1921_2020"

print("Clean Kaggle shape:", df.shape)
df.head()


df = df.dropna(subset=["track_id"])
df["track_id"] = df["track_id"].astype(str).str.strip()
df = df[df["track_id"] != ""]

df = df.drop_duplicates(subset="track_id").reset_index(drop=True)

print("Shape dopo deduplica:", df.shape)

# Convertiamo release_date in anno (utile per modello)
def extract_year(x):
    try:
        return int(str(x)[:4])
    except:
        return np.nan

df["year"] = df["release_date"].apply(extract_year)

# Riordino colonne
cols_new_order = [
    "track_id", "track_name", "artist_name", "popularity", "year",
    "acousticness", "danceability", "energy", "instrumentalness", "liveness",
    "loudness", "speechiness", "tempo", "valence", "duration_ms"
]

df = df[cols_new_order]
df.head()


output_path = os.path.join(DATA_PROCESSED_DIR, "spotify_dataset_clean.csv")
df.to_csv(output_path, index=False)

print("🎉 Dataset pulito salvato in:")
print(output_path)
